+++ DrawPool/Controls/PoolView.xaml.cs
@@ -25,39 +25,6 @@ namespace DrawPool.Controls
             InitializeComponent();
         }
 
-        public void Grid_MouseLeftButtonDown_UserControl(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        public void Grid_MouseLeftButtonDown_OuterBorder(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        public void Grid_MouseLeftButtonDown_InnerBorder(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        public void Grid_MouseLeftButtonDown_Panel(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        public void Grid_MouseLeftButtonDown_TextBlock(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        private void MD(object sender, MouseButtonEventArgs e)
-        {
-
-        }
-
-        private void MLBD(object sender, MouseButtonEventArgs e)
-        {
-
-        }
+      
     }
 }
