+++ DrawPool/DrawPoolPlugin.cs
@@ -1,12 +1,12 @@
 ﻿namespace DrawPool
 {
+    using DrawPool.Logic;
     using global::DrawPool.Controls;
     using Hearthstone_Deck_Tracker.Plugins;
     using System;
     using System.Reflection;
     using System.Windows.Controls;
     using Settings = Properties.Settings;
-    using Strings = Logic.Strings;
 
     /// <summary>
     /// The DrawPool plug-in Instance
@@ -14,13 +14,7 @@
     /// <seealso cref="Hearthstone_Deck_Tracker.Plugins.IPlugin" />
     public class DrawPoolPlugin : IPlugin
     {
-        
         internal DrawPoolInstance drawPoolInstance;
-
-        /// <summary>
-        /// The <see cref="DrawPoolWindow">DrawPool Window</see> reference
-        /// </summary>
-       // private DrawPoolWindow win;
 
         /// <summary>
         /// The author.
@@ -32,13 +26,13 @@
         /// The button text.
         /// </summary>
         /// <value>The button text.</value>
-        public string ButtonText => Strings.GetLocalized("SettingsLabel");
+        public string ButtonText => StringTools.GetLocalized("SettingsLabel");
 
         /// <summary>
         /// The plug-in description.
         /// </summary>
         /// <value>The plug-in description.</value>
-        public string Description => Strings.GetLocalized("PluginDescription");
+        public string Description => StringTools.GetLocalized("PluginDescription");
 
         /// <summary>
         /// Gets or sets the main <see cref="MenuItem">Menu Item</see>.
@@ -50,7 +44,7 @@
         /// The plug-in name.
         /// </summary>
         /// <value>The plug-in name.</value>
-        public string Name => Strings.GetLocalized("PluginName");
+        public string Name => StringTools.GetLocalized("PluginName");
 
         /// <summary>
         /// The plugin version.
@@ -97,7 +91,7 @@
 
         public void OnLoad()
         {
-           drawPoolInstance = new DrawPoolInstance();
+            drawPoolInstance = new DrawPoolInstance();
 
             AddMenuItem();
         }
@@ -118,7 +112,5 @@
         /// </summary>
         public void OnUpdate()
         { }
-
-
     }
 }
