+++ DrawPool/Logic/StringTools.cs
@@ -1,0 +1,17 @@
+﻿using WPFLocalizeExtension.Engine;
+
+namespace DrawPool.Logic
+{
+    /// <summary>
+    /// General String manipulation Tools
+    /// </summary>
+    public class StringTools
+    {
+        /// <summary>
+        /// Gets the localized version of the given string.
+        /// </summary>
+        /// <param name="key">The key.</param>
+        /// <returns>The localized string</returns>
+        public static string GetLocalized(string key) => LocalizeDictionary.Instance.GetLocalizedObject("DrawPool", "DrawPoolStrings", key, LocalizeDictionary.Instance.Culture)?.ToString();
+    }
+}
