--- DrawPool/Logic/Strings.cs
@@ -1,9 +1,0 @@
-﻿using WPFLocalizeExtension.Engine;
-
-namespace DrawPool.Logic
-{
-    public class Strings
-    {
-        public static string GetLocalized(string key) => LocalizeDictionary.Instance.GetLocalizedObject("DrawPool", "Strings", key, LocalizeDictionary.Instance.Culture)?.ToString();
-    }
-}
